+++ src/main/java/com/example/demo/client_ui/dto/order/OrderBriefDTO.java
@@ -15,7 +15,7 @@ public class OrderBriefDTO {
 
     private Integer numberOfItems;
 
-    private Double totalPrice;
+    private Float totalPrice;
 
     private String status;
 }
