+++ src/main/java/com/example/demo/client_ui/dto/product/ProductBriefDTO.java
@@ -5,11 +5,11 @@ import lombok.Data;
 @Data
 public class ProductBriefDTO {
 
-    private Long id;
+    private Integer id;
 
     private String name;
 
     private String imageUrl;
 
-    private Double price;
+    private Float price;
 }
